server/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Cors
python-dotenv
psycopg2-binary
Flask-JWT-Extended